Extract `TestTextResourcePropertiesService` to its own file
